+++ src/dotnet-serve/IPAddressParser.cs
@@ -1,0 +1,27 @@
+// Copyright (c) Nate McMaster.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System;
+using System.Globalization;
+using System.Net;
+using McMaster.Extensions.CommandLineUtils.Abstractions;
+
+namespace McMaster.DotNet.Server
+{
+    class IPAddressParser : IValueParser<IPAddress>
+    {
+        public Type TargetType => typeof(IPAddress);
+
+        public IPAddress Parse(string argName, string value, CultureInfo culture)
+        {
+            if (!IPAddress.TryParse(value, out var address))
+            {
+                throw new FormatException($"'{value}' is not a valid IP address");
+            }
+            return address;
+        }
+
+        object IValueParser.Parse(string argName, string value, CultureInfo culture)
+            => Parse(argName, value, culture);
+    }
+}