+++ src/dotnet-serve/RazorPages/RazorPageSourceProvider.cs
@@ -1,3 +1,6 @@
+// Copyright (c) Nate McMaster.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
 using System;
 using System.Collections.Concurrent;
 using System.IO;
@@ -19,13 +22,6 @@ namespace McMaster.DotNet.Server.RazorPages
         private readonly CSharpParseOptions _parseOptions;
         private ConcurrentDictionary<string, SyntaxTree> _syntaxTrees;
         private readonly FileSystemWatcher _fileWatcher;
-        private static Action<ILogger, string, Exception> _readingCsharpFile
-            = LoggerMessage.Define<string>(LogLevel.Debug, 0, "Parsing {file} as C#");
-
-        private static Action<ILogger, string, Exception> _csharpFileChanged
-            = LoggerMessage.Define<string>(LogLevel.Information, 0, "C# file changed: {file}");
-        private static Action<ILogger, string, int, int, string, string, Exception> _csharpSyntaxError
-            = LoggerMessage.Define<string, int, int, string, string>(LogLevel.Error, 0, "{file}({line},{col}): error {code}: {message}");
 
         public RazorPageSourceProvider(IHostingEnvironment env, ILoggerFactory loggerFactory)
         {
@@ -34,8 +30,6 @@ namespace McMaster.DotNet.Server.RazorPages
             _parseOptions = new CSharpParseOptions(LanguageVersion.Latest);
             _syntaxTrees = new ConcurrentDictionary<string, SyntaxTree>(StringComparer.Ordinal);
             _fileWatcher = new FileSystemWatcher(_sourceRoot, "*.cs");
-            ConfigureCompilerWatcher();
-            Initialize();
         }
 
         public void Initialize()
@@ -44,6 +38,7 @@ namespace McMaster.DotNet.Server.RazorPages
             {
                 ParseFile(file);
             }
+            ConfigureCompilerWatcher();
         }
 
         public void OnCompilation(RoslynCompilationContext ctx)
@@ -56,7 +51,10 @@ namespace McMaster.DotNet.Server.RazorPages
 
         private void ParseFile(string path)
         {
-            _readingCsharpFile(_logger, path, null);
+            if (_logger.IsEnabled(LogLevel.Trace))
+            {
+                _logger.LogTrace("Parsing {file} as C#", path);
+            }
 
             var source = SourceText.From(File.ReadAllText(path, Encoding.UTF8), Encoding.UTF8);
             var syntaxTree = CSharpSyntaxTree.ParseText(source, _parseOptions, path);
@@ -66,7 +64,7 @@ namespace McMaster.DotNet.Server.RazorPages
                 if (diag.Severity == DiagnosticSeverity.Error)
                 {
                     var pos = diag.Location.GetMappedLineSpan().StartLinePosition;
-                    _csharpSyntaxError(_logger, path, pos.Line, pos.Character, diag.Id, diag.GetMessage(), null);
+                    _logger.LogError("{file}({line},{col}): error {code}: {message}", _logger, path, pos.Line, pos.Character, diag.Id, diag.GetMessage());
                 }
             }
 
@@ -79,22 +77,34 @@ namespace McMaster.DotNet.Server.RazorPages
             _fileWatcher.Deleted += (o, e) =>
             {
                 _syntaxTrees.TryRemove(e.FullPath, out _);
-                Console.WriteLine("Detected file deletion " + e.FullPath);
+                if (_logger.IsEnabled(LogLevel.Debug))
+                {
+                    _logger.LogDebug("Detected file deletion of {path}", e.FullPath);
+                }
             };
             _fileWatcher.Renamed += (o, e) =>
             {
                 _syntaxTrees.TryRemove(e.OldFullPath, out _);
-                Console.WriteLine($"Detected rename {e.OldFullPath} => {e.FullPath}");
+                if (_logger.IsEnabled(LogLevel.Debug))
+                {
+                    _logger.LogDebug("Detected rename {old} => {new}", e.OldFullPath, e.FullPath);
+                }
                 ParseFile(e.FullPath);
             };
             _fileWatcher.Created += (o, e) =>
             {
-                Console.WriteLine("Detected new .cs file in " + e.FullPath);
+                if (_logger.IsEnabled(LogLevel.Debug))
+                {
+                    _logger.LogDebug("Detected new .cs file in {path}", e.FullPath);
+                }
                 ParseFile(e.FullPath);
             };
             _fileWatcher.Changed += (o, e) =>
             {
-                _csharpFileChanged(_logger, e.FullPath, null);
+                if (_logger.IsEnabled(LogLevel.Debug))
+                {
+                    _logger.LogDebug("C# file changed: {path}", e.FullPath);
+                }
                 ParseFile(e.FullPath);
             };
             _fileWatcher.EnableRaisingEvents = true;