+++ src/dotnet-serve/Startup.cs
@@ -1,40 +1,46 @@
+// Copyright (c) Nate McMaster.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System.Diagnostics;
 using McMaster.DotNet.Server.RazorPages;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.DataProtection.KeyManagement;
-using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace McMaster.DotNet.Server
 {
     class Startup
     {
+        private readonly CommandLineOptions _options;
+        private readonly RazorPageSourceProvider _sourceProvider;
+
         public Startup(
-            IConfiguration config,
+            CommandLineOptions options,
             RazorPageSourceProvider sourceProvider)
         {
-            Config = config;
+            _options = options;
             _sourceProvider = sourceProvider;
         }
-
-        IConfiguration Config { get; }
-
-        private readonly RazorPageSourceProvider _sourceProvider;
 
         public void ConfigureServices(IServiceCollection services)
         {
             services
-                .AddMvcCore()
-                    .AddRazorPages()
-                    .WithRazorPagesRoot("/")
-                    .AddRazorViewEngine(o =>
-                    {
-                        o.CompilationCallback = _sourceProvider.OnCompilation;
-                    })
-                    .Services
                 .Configure<KeyManagementOptions>(o => o.XmlRepository = new EphemeralXmlRepository())
                 .AddSingleton<IKeyManager, KeyManager>()
                 .AddSingleton<IAuthorizationPolicyProvider, NullAuthPolicyProvider>();
+
+            if (_options.EnableRazor)
+            {
+                Debug.Assert(_sourceProvider != null, "Source provider should be configured");
+                services.AddMvcCore()
+                      .AddRazorPages()
+                      .WithRazorPagesRoot("/")
+                      .AddRazorViewEngine(o =>
+                      {
+                          o.CompilationCallback = _sourceProvider.OnCompilation;
+                      });
+            }
         }
 
         public void Configure(IApplicationBuilder app)
@@ -43,11 +49,15 @@ namespace McMaster.DotNet.Server
                        "<html><head><title>Error {0}</title></head><body><h1>HTTP {0}</h1></body></html>");
 
             app.UseDeveloperExceptionPage();
-            app.UseMvcWithDefaultRoute();
-
-            if (!string.IsNullOrEmpty(Config["DotNetServe:PathBase"]))
+            if (_options.EnableRazor)
             {
-                app.Map(Config["DotNetServe:PathBase"], ConfigureFileServer);
+                app.UseMvcWithDefaultRoute();
+                _sourceProvider.Initialize();
+            }
+
+            if (!string.IsNullOrEmpty(_options.PathBase))
+            {
+                app.Map(_options.PathBase, ConfigureFileServer);
             }
             else
             {